refactor: update new DefaultIcon component imports
